feat: update p1 snapshot reactor; separate tests into examples and fixtures
