+++ Cargo.toml
@@ -12,10 +12,10 @@ crate-type = ["cdylib", "rlib"]
 members = ["./tests"]
 
 [workspace.dependencies]
-anyhow = "1.0.102"
+anyhow = "1.0.103"
 bzip2 = "0.6.1"
 once_cell = "1.21.3"
-reqwest = { version = "0.13.3", features = ["blocking"] }
+reqwest = { version = "0.13.4", features = ["blocking"] }
 tar = "0.4.46"
 
 [workspace.package]
@@ -43,11 +43,11 @@ bzip2 = { workspace = true }
 reqwest = { workspace = true }
 tar = { workspace = true }
 clap = { version = "4.6.1", features = ["derive"] }
-regex = "1.12.3"
+regex = "1.12.4"
 serde = { version = "1.0.228", features = ["derive"] }
 toml = "1.1.0"
-wit-bindgen-go = { git = "https://github.com/bytecodealliance/wit-bindgen", rev = "95995ce0abdfec2609d44d04537fe5378c1349c0" }
-wit-component = "0.249.0"
-wit-parser = "0.249.0"
-which = "8.0.2"
+wit-bindgen-go = { git = "https://github.com/asteurer/wit-bindgen", rev = "ea8159cf8cd2125d731ac72f8f3e0c64ca4e1d23" }
+wit-component = "0.252.0"
+wit-parser = "0.252.0"
+which = "8.0.4"
 dirs = "6.0.0"
