+++ tests/fixtures/multiple-worlds/export_wasi_http_handler/handler.go
