+++ tests/fixtures/multiple-worlds/export_wit_world/foo.go
