+++ tests/fixtures/multiple-worlds/wit1/deps/cli.wit
