+++ tests/fixtures/multiple-worlds/wit1/deps/clocks.wit
