+++ tests/fixtures/multiple-worlds/wit1/deps/http.wit
