+++ tests/fixtures/multiple-worlds/wit1/deps/sockets.wit
